+++ local.js
@@ -1824,6 +1824,7 @@ function extractUppercaseKeys(/*mutable*/ org, /*mutable*/ dst) {
 // Interface for receiving responses
 // - usually created internally and returned by `dispatch`
 function Response() {
+	var self = this;
 	local.util.EventEmitter.call(this);
 
 	this.status = 0;
@@ -1851,6 +1852,20 @@ function Response() {
 		writable: true
 	});
 
+	// header mixin
+	this.on('headers', function() {
+		for (var k in self.headers) {
+			var k2 = titlecaseHeader(k);
+			if (typeof self[k2] == 'undefined') {
+				Object.defineProperty(self, k2, { value: self.headers[k], configurable: true, enumerable: false, writable: true });
+			}
+			var k3 = underscorifyHeader(k2);
+			if (typeof self[k3] == 'undefined') {
+				Object.defineProperty(self, k3, { value: self.headers[k], configurable: true, enumerable: false, writable: true });
+			}
+		}
+	});
+
 	// response buffering
 	Object.defineProperty(this, 'body_', {
 		value: local.promise(),
@@ -1858,19 +1873,17 @@ function Response() {
 		enumerable: false,
 		writable: false
 	});
-	(function buffer(self) {
-		self.on('data', function(data) {
-			if (data instanceof ArrayBuffer)
-				self.body = data; // browsers buffer binary responses, so dont try to stream
-			else
-				self.body += data;
-		});
-		self.on('end', function() {
-			if (self.headers['content-type'])
-				self.body = local.contentTypes.deserialize(self.headers['content-type'], self.body);
-			self.body_.fulfill(self.body);
-		});
-	})(this);
+	this.on('data', function(data) {
+		if (data instanceof ArrayBuffer)
+			self.body = data; // browsers buffer binary responses, so dont try to stream
+		else
+			self.body += data;
+	});
+	this.on('end', function() {
+		if (self.headers['content-type'])
+			self.body = local.contentTypes.deserialize(self.headers['content-type'], self.body);
+		self.body_.fulfill(self.body);
+	});
 }
 local.Response = Response;
 Response.prototype = Object.create(local.util.EventEmitter.prototype);
@@ -1959,7 +1972,19 @@ Response.prototype.close = function() {
 	// this.removeAllListeners('end');
 	// this.removeAllListeners('close');
 	return this;
-};// Server
+};
+
+// internal helper
+var titlecaseRegExp = /(^(.))|(\-(.))/g;
+function titlecaseHeader(v) {
+	return v.replace(titlecaseRegExp, function(v) { return v.toUpperCase(); });
+}
+
+// internal helper
+var dashRegExp = /\-/g;
+function underscorifyHeader(v) {
+	return v.replace(dashRegExp, '_');
+}// Server
 // ======
 // EXPORTED
 // core type for all servers
@@ -5698,7 +5723,7 @@ function unbindRequestEvents(container) {
 // transforms click events into request events
 function Local__clickHandler(e) {
 	if (e.button !== 0) { return; } // handle left-click only
-	var request = local.util.extractRequest.fromAnchor(e.target);
+	var request = local.util.extractRequest.fromAnchor(e.orgtarget || e.target);
 	if (request && ['_top','_blank'].indexOf(request.target) !== -1) { return; }
 	if (request) {
 		e.preventDefault();